api/requirements.txt
Flask==3.0.0
kafka-python==2.0.2
mysql-connector-python==8.3.0